fixed big bug in the CodeGenerator.
Basically storing things in a std::vector and then getting a pointer or
reference to them is a big no-no, when the memory is reallocated because
the vector runs out of space, the pointer is dangling.

The solution is to have "ragged arrays" which hold unique_ptrs. Not the
most elegant, but works.